Base Activity for toolbar and global functions between the activities
