+++ app/src/main/res/layout/activity_settings_page.xml
@@ -19,4 +19,11 @@
         android:layout_alignParentBottom="true"
         android:layout_centerHorizontal="true"
         android:layout_marginBottom="228dp" />
+
+    <include layout="@layout/bottom_toolbar"
+        android:id="@+id/toolbar"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentStart="true"></include>
 </RelativeLayout>
