+++ app/src/main/res/layout/profile_overlay.xml
@@ -9,10 +9,10 @@
     >
 
     <ImageView
-        android:layout_width="200dp"
-        android:layout_height="200dp"
+        android:layout_width="156dp"
+        android:layout_height="147dp"
         app:srcCompat="@drawable/common_google_signin_btn_icon_dark_normal"
         android:id="@+id/my_qr"
-        android:layout_weight="0.33" />
+        android:layout_marginLeft="100dp" />
 
 </LinearLayout>
